refactored app for jsx and scalability
