+++ src/App.js
@@ -1,9 +1,9 @@
 import './App.css';
 import React from 'react';
-import Navbar from './components/navbarComponent';
+import Navbar from './pages/navbar/navbarComponent';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import { Container } from 'semantic-ui-react';
-
+import Main from './pages/primaryScreenshot/primaryInputPage'
 const Home = () => <h1>Home Page</h1>;
 const About = () => <h1>About Us</h1>;
 const Services = () => <h1>Our Services</h1>;
@@ -19,7 +19,7 @@ function App() {
         <Navbar authorised={isAuth} />
         <Container>
           <Routes>
-            <Route path="/" element={<Home />} />
+            <Route path="/" element={<Main authorised={isAuth}/>} />
             <Route path="/about" element={<About />} />
             <Route path="/services" element={<Services />} />
             <Route path="/contact" element={<Contact />} />
